Base commit after R&D, CUDA trial, Modular Code - NOT FUNCTIONAL YET
This is going to be the production ready version. 
Added modules and incorporated CUDA code in the required context. However the code is still not fully tested or running because of Memory Error when trying to store all Correlations in a bigger matrix. I guess I will need to put them by parts in 32x32 files each containing 27889x27889 matrix of correlation OR use only a triangular matrix to reduce memory requirements.

Once functional and debugged, this will complete the FS Enhancement Step which includes meta gene generation and clustering of the geneset.

Next steps are:
1) to use this enhanced dataset to train classifiers using LDA and calculate the J(.) Score
2) Form an ensemble of LDA's using Thinning algorithm
3) Write Classifier
4) Validate the system and compare results